+++ e2e/runner/ton.go
@@ -1,0 +1,59 @@
+package runner
+
+import (
+	"cosmossdk.io/math"
+	eth "github.com/ethereum/go-ethereum/common"
+	"github.com/stretchr/testify/require"
+	"github.com/tonkeeper/tongo/wallet"
+
+	toncontracts "github.com/zeta-chain/node/pkg/contracts/ton"
+)
+
+// we need to use this send mode due to how wallet V5 works
+//
+//	https://github.com/tonkeeper/w5/blob/main/contracts/wallet_v5.fc#L82
+//	https://docs.ton.org/develop/smart-contracts/guidelines/message-modes-cookbook
+const tonDepositSendCode = toncontracts.SendFlagSeparateFees + toncontracts.SendFlagIgnoreErrors
+
+// TONDeposit deposit TON to Gateway contract
+func (r *E2ERunner) TONDeposit(sender *wallet.Wallet, amount math.Uint, zevmRecipient eth.Address) error {
+	require.NotNil(r, r.TONGateway, "TON Gateway is not initialized")
+
+	require.NotNil(r, sender, "Sender wallet is nil")
+	require.False(r, amount.IsZero())
+	require.NotEqual(r, (eth.Address{}).String(), zevmRecipient.String())
+
+	r.Logger.Info(
+		"Sending deposit of %s TON from %s to zEVM %s",
+		amount.String(),
+		sender.GetAddress().ToRaw(),
+		zevmRecipient.Hex(),
+	)
+
+	return r.TONGateway.SendDeposit(r.Ctx, sender, amount, zevmRecipient, tonDepositSendCode)
+}
+
+// TONDepositAndCall deposit TON to Gateway contract with call data.
+func (r *E2ERunner) TONDepositAndCall(
+	sender *wallet.Wallet,
+	amount math.Uint,
+	zevmRecipient eth.Address,
+	callData []byte,
+) error {
+	require.NotNil(r, r.TONGateway, "TON Gateway is not initialized")
+
+	require.NotNil(r, sender, "Sender wallet is nil")
+	require.False(r, amount.IsZero())
+	require.NotEqual(r, (eth.Address{}).String(), zevmRecipient.String())
+	require.NotEmpty(r, callData)
+
+	r.Logger.Info(
+		"Sending deposit of %s TON from %s to zEVM %s and calling contract with %q",
+		amount.String(),
+		sender.GetAddress().ToRaw(),
+		zevmRecipient.Hex(),
+		string(callData),
+	)
+
+	return r.TONGateway.SendDepositAndCall(r.Ctx, sender, amount, zevmRecipient, callData, tonDepositSendCode)
+}